requirements.txt
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.0
scikit-learn==1.2.2
joblib==1.4.2
boto3==1.34.101
botocore==1.34.101
PyYAML==5.1.1
requests==2.31.0
pytest==7.2.0
mock==5.1.0